entry_fetch cleanup
