+++ script/csv_import/open_corporates/open_corporates_import_item_compact.rb
@@ -1,5 +1,6 @@
 require_relative "open_corporates_import_item"
 
+# import OC mappings (compact form)
 class OpenCorporatesImportItemCompact < OpenCorporatesImportItem
   @columns =
     [:oc_jurisdiction_code, :oc_company_number, :inc_jurisdiction_code,
